widget/Fader: Fix incorrect painter offset in Fader::paintEvent()
I'm currently seeing a segfault, caused by an incorrect target rectangle
offset generated by Fader::paintEvent.

If a Fader is created with type == PK_VU, and !WithoutKnob, then the
calculation for the x coordinate for painting the knob pixmap goes
negative (since the widget width is 11, and the knob pixmap is 15).
However, we're doing unsigned arithmetic here, so we end up with a huge
x coord, resulting in a segv in libQtGui.

This change uses signed arithmetic for the x-coordinate calculation, and
fixes the crash.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>