backend: fix crash when removing channels
Currently, I get a crash when removing channels or emptying the channels
table:

*** glibc detected *** /home/jk/build/livemix/livemix.git/livemix: double free or corruption (out): 0x0000000000931380 ***

The backtrace points to channel::~channel. This function calls erase()
on an iterator from the end of the list (it = effects.end()). However,
end() returns an iterator that is past the last element of the list, so
we shouldn't call erase() on this.

This change iterates from the beginning instead.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>